+++ src/core/client/stream/components/Comment/Timestamp.tsx
@@ -3,14 +3,14 @@ import { StatelessComponent } from "react";
 
 import { RelativeTime } from "talk-ui/components";
 
-import * as styles from "./CommentTimestamp.css";
+import * as styles from "./Timestamp.css";
 
-export interface CommentTimestampProps {
+export interface TimestampProps {
   children: string;
 }
 
-const CommentTimestamp: StatelessComponent<CommentTimestampProps> = props => (
+const Timestamp: StatelessComponent<TimestampProps> = props => (
   <RelativeTime className={styles.root} date={props.children} />
 );
 
-export default CommentTimestamp;
+export default Timestamp;
