+++ src/core/client/stream/components/Comment/index.ts
@@ -1,0 +1,1 @@
+export { default, default as Comment, CommentProps } from "./Comment";
